AlertManager/requirements.txt
fastapi
uvicorn
pyyaml
xmltodict
jsonschema
sqlalchemy
psycopg2-binary
geoalchemy2
